requirements.txt
langchain
langchain-core
langchain-community
faiss-cpu
langchain-nvidia-ai-endpoints
python-dotenv